MySQL Fixes
Started rewriting the MySQL data provider to fix the bugs I uncovered.

Rewriting the entire data provider around the idea that the server can
go away, and this corrupts prepared statements.

All the functionality the server actually currently uses is implemented,
but it's not back to where it was in terms of functionality.

Fixed/cleaned up the ping module.  100% working now as best as I can
tell.

Cleaned up some fluff in ThreadPool and NBT::Tag.