Make and Start of Chat Support
Started to implement a chat interface.  The chat module is itself going
to load modules, which will allow for chat to be modular and modifiable,
which is great since it's a part of the server so ripe for modification.

Made a few changes/tweaks.  Fixed a two's complement calculation bug
(carry wasn't being propagated) in the SHA-1 implementation.  Changed
the interface of the send functions in Client so they take a packet by
const reference, which allows for slightly more efficient chat
broadcasts (since the same Packet object can be reused).

Decided to switch over to make rather than a bitch script for building,
since the build times were getting long and I was tired of sitting
around waiting for the whole server to rebuild so that I could test one
or two minor changes in a module.

The old batch build script is still in the repo, but that's likely to
change soon.