Housekeeping
Working towards getting a successful Linux build so I can test the new
Linux network stack, no success, but progress.  Linux makefile is
progressing nicely, will have to merge some of the improvements I've
made to it back into the Windows makefile once I'm done.

Cleaned up the paths various #include directives were using to reflect
the realities of those libraries as they're installed on Linux, and
modified the include paths in the Windows build where possible to mirror
those.  Wasn't possible in the case of MySQL, but was in the case of
zlib, cURL, and OpenSSL.

Made a few slight changes to the network stack.

Added the ability to query Client/Connection objects as to the number of
pending send operations.

Apparently cURL on Linux (or, at least, the version I'm using) doesn't
have the CURLMOPT_PIPELINING option, so I added preprocessor directives
to remove it where it's not defined.