World Info Through Chat
Added a module which allows information about the WorldContainer to be
retrieved by typing "/info world" in chat (when you are an op).

Added the necessary hooks to support this functionality to the
WorldContainer.  Namely:

1. A function (and corresponding classes) to return all information
about the WorldContainer.
2. Functions which support the retrieval of human readable names for
dimensions on a per-world type basis.

Moreover, save time/count will now be recorded.

Fixed an omission in the ban info module -- a variable that should've
been marked static wasn't.