ColumnContainer Layout and Layout Verification
Added a "Populated" field to the ColumnContainer class -- this will be used to transmit whether the column is merely generated or also populated through saves and loads.

Added static asserts to verify the in memory layout of the Block class and the ColumnContainer class.  This allows the ColumnContainer's first three fields to be copied directly from memory to the backing store, eliminating the need for any intermediary copy.

The asserts also verify that the Block class does not have any alignment padding which means that the bare minimum of space in the backing store is used and that no undefined uninitialized memory is copied/read.